add .jshintrc so `npm test` won't fail
In the absence of a .jshintrc, jshint will use one from user's home directory which could cause jshint to fail, so jshint fails and accordingly `npm test` which calls jshint first.

So by creating an explicit .jshintrc we can prevent that it from using the wrong configuration.

It seems that the defaults for jshint@2.6.3 work fine for these files, so I just included an empty .jshintrc.

If there are specific jshintrc options that should also be enabled, they can be included in the .jshintrc file.